Use iterators for `for` loops
This introduces a few new Value types for Iterator and Range, and
various changes around the code to utilize Iterators. This was mainly
done to stop allocation of Arrays when creating a Range (`from..to`)
as this could be quite a huge memory burden with larger ranges. This
also gives some speed improvements.